+++ backend/controller/product/update-product-ctrl.js
@@ -1,9 +1,9 @@
 import Ajv from "ajv";
 
-import UPDATE_PRODUCT_SCHEMA from "../validator-schema/update-product-schema.js";
-import service from "../service/update-product-srvc.js";
+import UPDATE_PRODUCT_SCHEMA from "../../validator-schema/update-product-schema.js";
+import service from "../../service/update-product-srvc.js";
 
-const addProducrController = async (req, res) => {
+const addProductController = async (req, res) => {
   const ajv = new Ajv({ allErrors: true });
 
   const validate = ajv.compile(UPDATE_PRODUCT_SCHEMA);
@@ -25,4 +25,4 @@ const addProducrController = async (req, res) => {
   return await service(req, res);
 };
 
-export default addProducrController;
+export default addProductController;
