+++ backend/controller/product/view-product-ctrl.js
@@ -1,9 +1,9 @@
 import Ajv from "ajv";
 
-import VIEW_PRODUCT_SCHEMA from "../validator-schema/view-product-schema.js";
-import service from "../service/view-product-srvc.js";
+import VIEW_PRODUCT_SCHEMA from "../../validator-schema/view-product-schema.js";
+import service from "../../service/product/view-product-srvc.js";
 
-const viewProducrController = async (req, res) => {
+const viewProducController = async (req, res) => {
   const ajv = new Ajv({ allErrors: true });
 
   const validate = ajv.compile(VIEW_PRODUCT_SCHEMA);
@@ -15,4 +15,4 @@ const viewProducrController = async (req, res) => {
   return await service(req, res);
 };
 
-export default viewProducrController;
+export default viewProductController;
