+++ backend/utils/constants.js
@@ -1,5 +1,5 @@
 const INTERNAL_SERVER_ERROR = { status: 500, message: "Internal Server Error" };
 
-const PRODUCT_PER_PAGE = 10;
+const PRODUCT_PER_PAGE = 20;
 
 export { PRODUCT_PER_PAGE, INTERNAL_SERVER_ERROR };
